requirements.txt
python-dotenv
requests
pandas
SQLAlchemy
mysql-connector-python